requirements.txt
aiogram==3.24.0
aiohttp==3.13.3
python-dotenv==1.0.0
SQLAlchemy==2.0.36
aiosqlite==0.22.1
requests==2.32.5
